Fixing a bug in the ColorF class that caused color channels to come out on or off.

+++ ColorF.cs
@@ -82,7 +82,7 @@ namespace AsteroidOutpost
 		{
 			get
 			{
-				return new Color(R, G, B, A);
+				return new Color((int)R, (int)G, (int)B, (int)A);
 			}
 			set
 			{
